+++ CHANGES.rst
@@ -13,6 +13,22 @@ Changes
   The meta data us useful, for example, for selecting tests with the
   nose attribute mechanism.
 
+- Added ``doctestcase.doctestfiles``
+
+  - Define multiple doctest files at once.
+
+  - Automatically assign test class members.  So rather than::
+
+      class MYTests(unittest.TestCase):
+          ...
+          test_foo = doctestcase.doctestfile('foo.txt')
+
+    You can use::
+
+      @doctestcase.doctestfiles('foo.txt', 'bar.txt', ...)
+      class MYTests(unittest.TestCase):
+          ...
+
 4.4.0 (2015-07-16)
 ------------------